test: Use Mocha global fixtures for BrowserStack setup & teardown
